+++ T2-PTAmodel.C
@@ -51,8 +51,8 @@ double T2_PTAmodel(pulsar *psr,int p,int ipos,int param,int arr)
   longdouble omega;
   longdouble Sval,basicRoemer,x;
   longdouble pbdot;
-  longdouble deltaB=0.0L,dDB;
-  longdouble deltaSB=0.0L,model,r,s,deltaSR=0.0L,deltaAOP=0.0;
+  longdouble deltaB=longdouble(0.0),dDB;
+  longdouble deltaSB=longdouble(0.0),model,r,s,deltaSR=longdouble(0.0),deltaAOP=0.0;
   longdouble kom,kin,pmra,pmdec,dt0,Cval;
 	
 
@@ -66,9 +66,9 @@ double T2_PTAmodel(pulsar *psr,int p,int ipos,int param,int arr)
   }
   T0 = psr[p].param[param_t0].val[0];
   do {
-    tbbat = psr[p].obsn[ipos].bbat - deltaB/86400.0L;
+    tbbat = psr[p].obsn[ipos].bbat - deltaB/longdouble(86400.0);
     
-    dt = (tbbat - T0)*86400.0L;
+    dt = (tbbat - T0)*longdouble(86400.0);
     
     // Eccentricity
     if (psr[p].param[param_ecc].paramSet[0] == 0) {
@@ -78,7 +78,7 @@ double T2_PTAmodel(pulsar *psr,int p,int ipos,int param,int arr)
     if (psr[p].param[param_edot].paramSet[0] == 1) 
       edot = psr[p].param[param_edot].val[0];
     else
-      edot = 0.0L;
+      edot = longdouble(0.0);
     
     // Equation 64 in Edwards et al. 2006
     e = e0 + edot*dt;
@@ -87,10 +87,10 @@ double T2_PTAmodel(pulsar *psr,int p,int ipos,int param,int arr)
     if (psr[p].param[param_pb].paramSet[0] == 0) {
       printf("Must use PB in the binary model\n"); exit(1);
     }
-    pb = psr[p].param[param_pb].val[0]*86400.0L;
+    pb = psr[p].param[param_pb].val[0]*longdouble(86400.0);
     
-    nval = 2.0L*M_PI/pb; 
-    pbdot = 0.0L;
+    nval = longdouble(2.0)*M_PI/pb; 
+    pbdot = longdouble(0.0);
     if (psr[p].param[param_pbdot].paramSet[0]==1)
       {
 	pbdot = psr[p].param[param_pbdot].val[0];
@@ -100,10 +100,10 @@ double T2_PTAmodel(pulsar *psr,int p,int ipos,int param,int arr)
       }
 
     // HOW DOES THIS RELATE TO THE EDWARDS PAPER?
-    orbits = nval/2.0L/M_PI*dt;
+    orbits = nval/longdouble(2.0)/M_PI*dt;
     norbits = (int)(orbits);
-    if (orbits < 0.0L) norbits --;  // SHOULD CHECK THIS
-    phase = 2.0L*M_PI*(orbits-norbits);
+    if (orbits < longdouble(0.0)) norbits --;  // SHOULD CHECK THIS
+    phase = longdouble(2.0)*M_PI*(orbits-norbits);
     
     //    phase = nval*dt;
     //    printf("Phase = %g %g %g %g\n",(double)phase,(double)pb,(double)dt,(double)norbits);
@@ -112,8 +112,8 @@ double T2_PTAmodel(pulsar *psr,int p,int ipos,int param,int arr)
     //    printf("U = %g\n",(double)u);
 
     // Equation 58 -- should check how to calculate this atan - USE LONG DOUBLE MATHS?
-    //Ae = 2.0L*atanl(powl((1.0L+e)/(1.0L-e),2)*tanl(u/2.0L));
-    Ae = 2.0L*atan2((1.0L+e)*(1.0L+e)*tan(u/2.0),(1.0L-e)*(1.0L-e));
+    //Ae = longdouble(2.0)*atanl(powl((longdouble(1.0)+e)/(longdouble(1.0)-e),2)*tanl(u/longdouble(2.0)));
+    Ae = longdouble(2.0)*atan2((longdouble(1.0)+e)*(longdouble(1.0)+e)*tan(u/2.0),(longdouble(1.0)-e)*(longdouble(1.0)-e));
     
     //    Ae = atan2(sqrt(1.0-e*e)*sin(u)/(1.0-e*cos(u)),(cos(u)-e)/(1.0-e*cos(u)));
         if(Ae<0.0) Ae=Ae+2.0*M_PI;
@@ -123,19 +123,19 @@ double T2_PTAmodel(pulsar *psr,int p,int ipos,int param,int arr)
     
     // Equation 63
     if (psr[p].param[param_omdot].paramSet[0]==1){
-      kval = psr[p].param[param_omdot].val[0]*M_PI/180.0/(365.25*86400.0L)/nval;
+      kval = psr[p].param[param_omdot].val[0]*M_PI/180.0/(365.25*longdouble(86400.0))/nval;
     }
     else
-      kval = 0.0L;
+      kval = longdouble(0.0);
     
     // Equation 62
     if (psr[p].param[param_om].paramSet[0] == 0) {
       printf("Must use OM in the binary model\n"); exit(1);
     }
-    omega = (psr[p].param[param_om].val[0]*M_PI/180.0L+kval*Ae);
+    omega = (psr[p].param[param_om].val[0]*M_PI/longdouble(180.0)+kval*Ae);
     //    printf("omega %g %g %g %g %g\n",(double)psr[0].obsn[ipos].bbat,(double)omega,(double)Ae,(double)kval,(double)(kval*Ae));
     // Equation 65
-    Sval = sinl(omega)*(cosl(u)-e)+cosl(omega)*sqrt(1.0L-e*e)*sinl(u);
+    Sval = sinl(omega)*(cosl(u)-e)+cosl(omega)*sqrt(longdouble(1.0)-e*e)*sinl(u);
     
     
     // Equation 71
@@ -154,10 +154,10 @@ double T2_PTAmodel(pulsar *psr,int p,int ipos,int param,int arr)
     if (psr[p].param[param_kin].paramSet[0]==1 &&
 	psr[p].param[param_m2].paramSet[0]==1)
       {
-	kin = psr[p].param[param_kin].val[0]*M_PI/180.0L;
+	kin = psr[p].param[param_kin].val[0]*M_PI/longdouble(180.0);
 	s = sin(kin);
 	r = psr[p].param[param_m2].val[0]*SUNMASS;
-	deltaSB = -2*r*log(1.0L-e*cosl(u) - s*(sinl(omega)*(cosl(u)-e) + sqrt(1.0-e*e)*cos(omega)*sin(u)));
+	deltaSB = -2*r*log(longdouble(1.0)-e*cosl(u) - s*(sinl(omega)*(cosl(u)-e) + sqrt(1.0-e*e)*cos(omega)*sin(u)));
       }
 
     // Calculate Kopeikin terms
@@ -166,14 +166,14 @@ double T2_PTAmodel(pulsar *psr,int p,int ipos,int param,int arr)
 	psr[p].param[param_kin].paramSet[0]==1 &&
 	psr[p].param[param_kom].paramSet[0]==1)
       {
-	dt0 = (psr[p].obsn[ipos].bbat - psr[p].param[param_posepoch].val[0])*86400.0L;
+	dt0 = (psr[p].obsn[ipos].bbat - psr[p].param[param_posepoch].val[0])*longdouble(86400.0);
 
 	kin = psr[p].param[param_kin].val[0]*M_PI/180.0;
 	kom = psr[p].param[param_kom].val[0]*M_PI/180.0;
-	pmra = psr[p].param[param_pmra].val[0]*M_PI/(180.0*3600.0e3)/(365.25*86400.0L);
-	pmra = psr[p].param[param_pmdec].val[0]*M_PI/(180.0*3600.0e3)/(365.25*86400.0L);
+	pmra = psr[p].param[param_pmra].val[0]*M_PI/(180.0*3600.0e3)/(365.25*longdouble(86400.0));
+	pmra = psr[p].param[param_pmdec].val[0]*M_PI/(180.0*3600.0e3)/(365.25*longdouble(86400.0));
 
-	Cval = cosl(omega)*(cosl(u)-e)-sinl(omega)*sqrtl(1.0L-e*e)*sinl(u);
+	Cval = cosl(omega)*(cosl(u)-e)-sinl(omega)*sqrtl(longdouble(1.0)-e*e)*sinl(u);
 	
 	// Equation 73
 	deltaSR = x*dt0*((pmra*sin(kom)+pmdec*cos(kom))*Cval/sin(kin) + (pmra*cos(kom)-pmdec*sin(kom))*Sval/tan(kin));
@@ -222,29 +222,29 @@ double T2_PTAmodel(pulsar *psr,int p,int ipos,int param,int arr)
   longdouble domega_domegaDot,dD_dkom,dD_di,dD_dm2;
 
   // Not including time derivatives of anything yet
-  //  du_dpb = (-2.0L*M_PI*dt/pb/pb - 2.0*M_PI*dt*dt*pbdot/powl(pb,3))/(1.0L-e*cosl(u));
-  du_dpb = (-2.0L*M_PI*dt/pb/pb)/(1.0L-e*cosl(u));
-  du_dT0 = -2.0L*M_PI/pb/(1.0L-e*cosl(u));
-  du_de  = sinl(u)/(1.0L-e*cosl(u));
+  //  du_dpb = (-longdouble(2.0)*M_PI*dt/pb/pb - 2.0*M_PI*dt*dt*pbdot/powl(pb,3))/(longdouble(1.0)-e*cosl(u));
+  du_dpb = (-longdouble(2.0)*M_PI*dt/pb/pb)/(longdouble(1.0)-e*cosl(u));
+  du_dT0 = -longdouble(2.0)*M_PI/pb/(longdouble(1.0)-e*cosl(u));
+  du_de  = sinl(u)/(longdouble(1.0)-e*cosl(u));
   // -ve sign in DDmodel.C
-  du_dpdot = M_PI*dt*dt/pb/pb/(1.0L-e*cosl(u));
+  du_dpdot = M_PI*dt*dt/pb/pb/(longdouble(1.0)-e*cosl(u));
 
   // This seems odd
   domega_domegaDot = Ae/(nval*180.0/M_PI*365.25*SECDAY);
 
-  dD_da1 = sinl(omega)*(cosl(u)-e)+cosl(omega)*sqrt(1.0L-e*e)*sinl(u);
+  dD_da1 = sinl(omega)*(cosl(u)-e)+cosl(omega)*sqrt(longdouble(1.0)-e*e)*sinl(u);
   dD_de = -x*sinl(omega)*sinl(u)*du_de - x*sinl(omega) 
-    - x*cosl(omega)/sqrt(1.0L-e*e)*e*sinl(u) + x*cosl(omega)*sqrt(1-e*e)*cosl(u)*du_de;
-  dD_dT0 = (-x*sinl(omega)*sinl(u) + x*cosl(omega)*sqrt(1.0L-e*e)*cosl(u))*du_dT0;
-  dD_dpb = (-x*sinl(omega)*sinl(u) + x*cosl(omega)*sqrt(1.0L-e*e)*cosl(u))*du_dpb;
-  dD_dpdot = (-x*sinl(omega)*sinl(u) + x*cosl(omega)*sqrt(1.0L-e*e)*cosl(u))*du_dpdot;
-  dD_domega = x*(cosl(u)-e)*cosl(omega) - x*sqrt(1.0L-e*e)*sinl(u)*sinl(omega);
+    - x*cosl(omega)/sqrt(longdouble(1.0)-e*e)*e*sinl(u) + x*cosl(omega)*sqrt(1-e*e)*cosl(u)*du_de;
+  dD_dT0 = (-x*sinl(omega)*sinl(u) + x*cosl(omega)*sqrt(longdouble(1.0)-e*e)*cosl(u))*du_dT0;
+  dD_dpb = (-x*sinl(omega)*sinl(u) + x*cosl(omega)*sqrt(longdouble(1.0)-e*e)*cosl(u))*du_dpb;
+  dD_dpdot = (-x*sinl(omega)*sinl(u) + x*cosl(omega)*sqrt(longdouble(1.0)-e*e)*cosl(u))*du_dpdot;
+  dD_domega = x*(cosl(u)-e)*cosl(omega) - x*sqrt(longdouble(1.0)-e*e)*sinl(u)*sinl(omega);
   dD_domegaDot = dD_domega*domega_domegaDot;
 
 
   // Currently only including Shapiro term
   dD_dm2 = deltaSB/r;
-  dD_di = -2*r/(1.0L-e*cosl(u) - s*(sinl(omega)*(cosl(u)-e) + sqrt(1.0-e*e)*cos(omega)*sin(u)))*(-cos(kin)*(sinl(omega)*(cosl(u)-e) + sqrt(1.0-e*e)*cos(omega)*sin(u)));
+  dD_di = -2*r/(longdouble(1.0)-e*cosl(u) - s*(sinl(omega)*(cosl(u)-e) + sqrt(1.0-e*e)*cos(omega)*sin(u)))*(-cos(kin)*(sinl(omega)*(cosl(u)-e) + sqrt(1.0-e*e)*cos(omega)*sin(u)));
 
   // 
   if (param==param_kom)
@@ -273,7 +273,7 @@ double T2_PTAmodel(pulsar *psr,int p,int ipos,int param,int arr)
   else if (param==param_t0)
     {
       printf("dT0 = %g\n",(double)dD_dT0);
-      return dD_dT0*86400.0L;
+      return dD_dT0*longdouble(86400.0);
     }
   else if (param==param_ecc)
     return dD_de;
@@ -304,18 +304,18 @@ void updateT2_PTA(pulsar *psr,double val,double err,int pos,int arr){
     }
   else if (pos==param_om) // || pos==param_t0)
     {
-      psr->param[pos].val[0] += val*180.0L/M_PI;
-      psr->param[pos].err[0]  = err*180.0L/M_PI;
+      psr->param[pos].val[0] += val*longdouble(180.0)/M_PI;
+      psr->param[pos].err[0]  = err*longdouble(180.0)/M_PI;
     }
   else if (pos==param_kom) // || pos==param_t0)
     {
-      psr->param[pos].val[0] += val*180.0L/M_PI;
-      psr->param[pos].err[0]  = err*180.0L/M_PI;
+      psr->param[pos].val[0] += val*longdouble(180.0)/M_PI;
+      psr->param[pos].err[0]  = err*longdouble(180.0)/M_PI;
     }
   else if (pos==param_kin) 
     {
-      psr->param[pos].val[0] += val*180.0L/M_PI;
-      psr->param[pos].err[0]  = err*180.0L/M_PI;
+      psr->param[pos].val[0] += val*longdouble(180.0)/M_PI;
+      psr->param[pos].err[0]  = err*longdouble(180.0)/M_PI;
     }
   else if (pos==param_m2) 
     {
@@ -324,8 +324,8 @@ void updateT2_PTA(pulsar *psr,double val,double err,int pos,int arr){
     }
   else if (pos==param_omdot) 
     {
-      psr->param[pos].val[0] += val;//*180.0L/M_PI*86400.0*365.25;
-      psr->param[pos].err[0]  = err;//*180.0L/M_PI*86400.0*365.25;
+      psr->param[pos].val[0] += val;//*longdouble(180.0)/M_PI*86400.0*365.25;
+      psr->param[pos].err[0]  = err;//*longdouble(180.0)/M_PI*86400.0*365.25;
     }
   else 
     {
